should be a long, not an int
git-svn-id: https://svn.apache.org/repos/asf/incubator/abdera/java/trunk@629227 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/abdera/util/AbstractStreamWriter.java
@@ -127,7 +127,7 @@ public abstract class AbstractStreamWriter
     String type, 
     String title,
     String hreflang, 
-    int length) {
+    long length) {
       return 
         startElement(Constants.LINK).
         writeAttribute("href",iri).
@@ -345,7 +345,7 @@ public abstract class AbstractStreamWriter
     String type, 
     String title,
     String hreflang, 
-    int length) {
+    long length) {
       return 
         startLink(iri, rel, type, title, hreflang, length).
         endLink();
